feat: pull out EnvironmentConfigManager to its own library, change nodejs build to define external pacakges
