+++ lib/shared/config-manager/README.md
@@ -1,0 +1,11 @@
+# config-manager
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build config-manager` to build the library.
+
+## Running unit tests
+
+Run `nx test config-manager` to execute the unit tests via [Jest](https://jestjs.io).
