+++ lib/shared/config-manager/jest.config.ts
@@ -1,0 +1,13 @@
+/* eslint-disable */
+export default {
+    displayName: 'config-manager',
+    preset: '../../../jest.preset.js',
+    transform: {
+        '^.+\\.[tj]s$': [
+            'ts-jest',
+            { tsconfig: '<rootDir>/tsconfig.spec.json' },
+        ],
+    },
+    moduleFileExtensions: ['ts', 'js', 'html'],
+    coverageDirectory: '../../../coverage/lib/shared/config-manager',
+}
