+++ lib/shared/config-manager/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@devcycle/config-manager",
+  "version": "0.0.1",
+  "type": "commonjs"
+}
