+++ lib/shared/config-manager/src/lib/config-manager.spec.ts
@@ -1,0 +1,7 @@
+import { configManager } from './config-manager'
+
+describe('configManager', () => {
+    it('should work', () => {
+        expect(configManager()).toEqual('config-manager')
+    })
+})
