+++ lib/shared/config-manager/src/lib/config-manager.ts
@@ -1,0 +1,3 @@
+export function configManager(): string {
+    return 'config-manager'
+}
