+++ sdk/nodejs/__tests__/utils/setPlatformData.spec.ts
