+++ sdk/nodejs/package.json
@@ -1,15 +1,17 @@
 {
-    "name": "@devcycle/nodejs-server-sdk",
-    "version": "1.15.5",
-    "description": "The DevCycle NodeJS Server SDK used for feature management.",
-    "license": "MIT",
-    "devDependencies": {
-        "@devcycle/bucketing-test-data": "^1.2.22"
-    },
-    "dependencies": {
-        "@devcycle/bucketing-assembly-script": "^1.8.4"
-    },
-    "engines": {
-        "node": ">=14.0.0"
-    }
+  "name": "@devcycle/nodejs-server-sdk",
+  "version": "1.15.5",
+  "description": "The DevCycle NodeJS Server SDK used for feature management.",
+  "license": "MIT",
+  "devDependencies": {
+    "@devcycle/bucketing-test-data": "^1.2.22"
+  },
+  "dependencies": {
+    "@devcycle/bucketing-assembly-script": "^1.8.4",
+    "@devcycle/js-cloud-server-sdk": "^1.0.0",
+    "@devcycle/types": "^1.1.15"
+  },
+  "engines": {
+    "node": ">=14.0.0"
+  }
 }
