+++ sdk/nodejs/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "files": [],
     "include": [],
+    "paths": {},
     "references": [
         {
             "path": "./tsconfig.lib.json"
